Tidy up types and add return val to subscriber type
